add test and reorganize function in files
