+++ alntools/density/__init__.py
@@ -1,3 +1,3 @@
-from .local import get_multires_density, chunk_cosine_similarity, batch_slice_iterator, reduce_duplicates_query_filedict
+from .local import get_multires_density, chunk_cosine_similarity, batch_slice_iterator
 from .parallel import load_and_score_database
 from .parallel import load_full_embeddings
