+++ alntools/density/local.py
@@ -319,17 +319,6 @@ def chunk_cosine_similarity(query : Union[th.Tensor, List[th.Tensor]],
 	return results
 
 
-def reduce_duplicates_query_filedict(query_filedict):
-	for key, value in query_filedict.items():
-		for k in list(value.keys()):
-			if k < key and query_filedict.get(key).get(k):
-				del query_filedict[key][k]
-			if k >= key:
-				continue
-
-	return query_filedict
-
-
 @th.jit.script
 def norm_chunk(target, kernel_size: int, embdim: int, stride: int):
 	'''
