+++ alntools/prepare/__init__.py
@@ -1,1 +1,2 @@
 from .iterate import search_paths
+from .reduce_duplicates import reduce_duplicates_query_filedict
